% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODPATHmass.R
\name{MassTrack}
\alias{MassTrack}
\title{MassTrack}
\usage{
MassTrack(ptl = "pathline", gwnc, wtop = "wtop.nc", phi, m0 = 1,
  retain.storage = TRUE, truncate = TRUE, confined = FALSE,
  outflux = TRUE, loss = TRUE, outflux.array = TRUE,
  tlumpoutflux = FALSE, linear.decay = FALSE, react.loss = TRUE, end.t)
}
\arguments{
\item{ptl}{a data.frame, data.table or string giving the path to a data
file containing a pathline data set produced by MODPATH}

\item{gwnc}{a NetCDF of MODFLOW simulation data produced by GW.nc, or a
path to one}

\item{wtop}{a NetCDF of water height or path to one (MassTrack can
generate one if this hasn't been made)}

\item{phi}{porosity: single value, layer by layer (vector) or
distributed array}

\item{m0}{initial mass of particles: single value, vector of values or
function of time only}

\item{retain.storage}{should mass flux to storage be retained on a
pathline?}

\item{truncate}{if TRUE, once a particle is completely drained, further
trajectory information is not returned with the result}

\item{confined}{not yet implemented}

\item{outflux}{return cell referenced information about where the mass
is abstracted?}

\item{loss}{return information about where particle mass is not released}

\item{outflux.array}{logical \code{[1]};
if outflux = TRUE, then outflux.array = TRUE signals that the abstracted
 mass information should be returned in an array rather than a data.table}

\item{tlumpoutflux}{logical \code{[1]};
should the outflux array only show the spatial cell location (TRUE), or
 show during which time step the abstractions occurred as well (in which
 case the outflux array will be four-dimensional and potentially very
 large)}

\item{linear.decay}{double \code{[1]};
linear decay constant, defaults to 0 which is no decay}

\item{react.loss}{logical \code{[1]};
similar to outflux, save locational reaction loss information?}

\item{end.t}{double \code{[1]};
specify the end time of the simulation; if missing, it will be inferred
 from ptl, but the last particles to arrive at an abstraction may not be
 drained because the simulation will be assumed to have finished as soon
 as or just after they arrive}
}
\value{
A list of results:
$trace, data.table: the pathline data table with new columns for mass
 and, if \code{outflux.array = FALSE}, mass loss and reactive mass
 loss\cr
$outflux, numeric array: if \code{outflux = TRUE} and
 \code{outflux.array = TRUE}, the mass removed from the system by
 column, row and layer and, if \code{tlumpoutflux = FALSE}, time step\cr
$react.loss, numeric array: as with outflux but for reactive loss\cr
$loss, numeric \code{[Np]}: mass which is lost otherwise (e.g. particle
 wasn't released)

Only requested outputs are returned and if only traces are returned,
 then the data table is not put in a list.
}
\description{
Track particle mass along a MODPATH-5 generated pathline
}
\examples{
# to follow

}

